requirements.txt
# Core NLP & ML
sentence-transformers>=2.2.2
chromadb>=0.4.15
bertopic>=0.15.0
scikit-learn>=1.3.2
umap-learn>=0.5.4
hdbscan>=0.8.33
spacy>=3.7.2
nltk>=3.8.1
transformers>=5.0.0
torch>=2.1.0
setfit>=1.0.0

# Data Processing
pandas>=2.1.2
numpy>=1.26.0
tiktoken>=0.5.1
beautifulsoup4>=4.12.2
regex>=2023.10.3
pyarrow>=14.0.0

# LLM & API
openai>=1.2.0
tenacity>=8.2.3

# Backend API
fastapi>=0.104.1
uvicorn>=0.23.2
pydantic>=2.4.2
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.23

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0
wordcloud>=1.9.3
streamlit>=1.28.0

# Evaluation
rouge-score>=0.1.2
bert-score>=0.3.13

# Testing
pytest>=7.4.3
httpx>=0.25.0

# Utilities
requests>=2.31.0
tqdm>=4.66.1
loguru>=0.7.2
